Fix problems with RelayRouter server
The design of the RelayRouter is still not formed completely, but there
were a lot of problems that needed to be ammended.

The relay router connection pool was originally written to be in the
network library, but since the functionality is very specific to the
relay router (it's the only place that is using it currently) it should
just go, and stay in there. This removes all the connection pool
management from network.

The relay router communicates by a single channel of relay requests that
controls. This channel is read from in a goroutine that manages all of
the connections.